+++ CHANGELOG.md
@@ -10,3 +10,9 @@ The initial initial implementation was contributed on behalf of STMicroelectroni
 ### Changes
 
 -   [diagram] Implement LayoutEngine API for server-side autolayouting & provide an integration package for layout engines based on ELK. [#509](https://github.com/eclipse-glsp/glsp-server-node/pull/2) [#514](https://github.com/eclipse-glsp/glsp-server-node/pull/5) - Contributed on behalf of STMicroelectronics
+
+### Breaking Changes
+
+-   [server] Source model refactorings [#582](https://github.com/eclipse-glsp/glsp/issues/582)
+    -   `ModelSourceLoader` → `SourceModelStorage`
+    -   Added method to `SourceModelStorage`